+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [13.26.2](https://kaos.sh/ek/13.26.2)
 
 - **`[support]`** Improved handling empty slices with data (_deps, apps, services…_)
+- **`[support/apps]`** Added method `ExtractVersion` for extracting version info from command output
 
 ### [13.26.1](https://kaos.sh/ek/13.26.1)